inchaellah

+++ libft/ft_strncmp.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-size_t	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned char	*p1;
 	unsigned char	*p2;
