+++ libft/libft.h
@@ -36,7 +36,7 @@ char				*ft_strchr(const char *s, int c);
 char				*ft_strrchr(const char *s, int c);
 size_t				ft_strlcpy(char *dst, const char *src, size_t size);
 size_t				ft_strlcat(char *dst, const char *src, size_t size);
-size_t				ft_strncmp (const char *s1, const char *s2, size_t n);
+int					ft_strncmp (const char *s1, const char *s2, size_t n);
 char				*ft_strnstr(const char *heystack,
 					const char *needle, size_t len);
 void				ft_bzero(void *s, size_t n);
